+++ app/src/main/java/com/mikhalevich/roadtogoal/ChildrenGoalsActivity.java
@@ -1,29 +1,25 @@
 package com.mikhalevich.roadtogoal;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
 
-public class ChildrenGoalsActivity extends AppCompatActivity {
+import com.mikhalevich.roadtogoal.domain.dbtasks.LoadGoalsTask;
+
+import lombok.Getter;
+
+public class ChildrenGoalsActivity extends BasicActivity {
+    @Getter
+    private Integer parentId;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_children_goals);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
 
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-            }
-        });
+        parentId = getIntent().getIntExtra("parentId", -1);
+        if (parentId == -1)
+            throw new RuntimeException();
+        super.initializeFloatingActionButton(this, parentId);
+
+        LoadGoalsTask task = new LoadGoalsTask(super.goalList, super.recyclerView, parentId);
+        task.execute();
     }
-
 }
